[bug] fix int16 type typo, fangq/iso2mesh#83

+++ nii2jnii.m
@@ -177,7 +177,7 @@ type2str = {
             'uint8'    0   % unknown                       %
             'uint8'    0   % binary (1 bit/voxel)          %
             'uint8'    1   % unsigned char (8 bits/voxel)  %
-            'uint16'   1   % signed short (16 bits/voxel)  %
+            'int16'    1   % signed short (16 bits/voxel)  %
             'int32'    1   % signed int (32 bits/voxel)    %
             'single'   1   % float (32 bits/voxel)         %
             'single'   2   % complex (64 bits/voxel)       %
@@ -236,7 +236,7 @@ else
         nii.img = fread(fid, imgbytenum, [nii.datatype '=>' nii.datatype]);
         fclose(fid);
     else
-        nii.img = typecast(gzdata(nii.hdr.vox_offset + 1:nii.hdr.vox_offset + imgbytenum), nii.datatype);
+        nii.img = typecast(gzdata(double(nii.hdr.vox_offset + 1):double(nii.hdr.vox_offset + imgbytenum)), nii.datatype);
     end
 end
